+++ libbeat/common/transport/tlscommon/tls_config.go
@@ -1,0 +1,77 @@
+package tlscommon
+
+import (
+	"crypto/tls"
+	"crypto/x509"
+
+	"github.com/elastic/beats/libbeat/logp"
+)
+
+// TLSConfig is the interface used to configure a tcp client or server from a `Config`
+type TLSConfig struct {
+
+	// List of allowed SSL/TLS protocol versions. Connections might be dropped
+	// after handshake succeeded, if TLS version in use is not listed.
+	Versions []TLSVersion
+
+	// Configure SSL/TLS verification mode used during handshake. By default
+	// VerifyFull will be used.
+	Verification TLSVerificationMode
+
+	// List of certificate chains to present to the other side of the
+	// connection.
+	Certificates []tls.Certificate
+
+	// Set of root certificate authorities use to verify server certificates.
+	// If RootCAs is nil, TLS might use the system its root CA set (not supported
+	// on MS Windows).
+	RootCAs *x509.CertPool
+
+	// Set of root certificate authorities use to verify client certificates.
+	// If ClientCAs is nil, TLS might use the system its root CA set (not supported
+	// on MS Windows).
+	ClientCAs *x509.CertPool
+
+	// List of supported cipher suites. If nil, a default list provided by the
+	// implementation will be used.
+	CipherSuites []uint16
+
+	// Types of elliptic curves that will be used in an ECDHE handshake. If empty,
+	// the implementation will choose a default.
+	CurvePreferences []tls.CurveID
+
+	// Renegotiation controls what types of renegotiation are supported.
+	// The default, never, is correct for the vast majority of applications.
+	Renegotiation tls.RenegotiationSupport
+
+	// ClientAuth controls how we want to verify certificate from a client, `none`, `optional` and
+	// `required`, default to required. Do not affect TCP client.
+	ClientAuth tls.ClientAuthType
+}
+
+// BuildModuleConfig takes the TLSConfig and tranform it into a `tls.Config`.
+func (c *TLSConfig) BuildModuleConfig(host string) *tls.Config {
+	if c == nil {
+		// use default TLS settings, if config is empty.
+		return &tls.Config{ServerName: host}
+	}
+
+	minVersion, maxVersion := extractMinMaxVersion(c.Versions)
+	insecure := c.Verification != VerifyFull
+	if insecure {
+		logp.Warn("SSL/TLS verifications disabled.")
+	}
+
+	return &tls.Config{
+		ServerName:         host,
+		MinVersion:         minVersion,
+		MaxVersion:         maxVersion,
+		Certificates:       c.Certificates,
+		RootCAs:            c.RootCAs,
+		ClientCAs:          c.ClientCAs,
+		InsecureSkipVerify: insecure,
+		CipherSuites:       c.CipherSuites,
+		CurvePreferences:   c.CurvePreferences,
+		ClientAuth:         c.ClientAuth,
+	}
+}
